Fix TUI compilation by restoring ui/mod.rs and adding missing domain types
This commit resolves ~28 TUI compilation errors by:

1. RESTORE ui/mod.rs as module entry point (was deleted, replacing the 9-tab
   live module with the 5-tab ui.rs). Rename ui.rs -> ui_live.rs to avoid
   Rust module ambiguity (ui.rs shadows ui/mod.rs when both exist).

2. api: Add BenchmarksResponse (SofrBenchmarksResponse + TreasuryBenchmarksResponse),
   add Deserialize to CurveResponse, SofrBenchmarksResponse,
   SofrOvernightResponse, TreasuryBenchmarksResponse, BenchmarkRateResponse.
   Add ScenarioDto with all fields needed by scenarios tab (symbol, expiration,
   strike_width, strike_center, days_to_expiry, theoretical_value,
   estimated_net_debit, net_debit, profit, roi_pct, apr_pct, fill_probability,
   scenario_type).

3. api: Add scenarios: Vec<ScenarioDto> and yield_benchmarks:
   Option<BenchmarksResponse> to RuntimeSnapshotDto.

4. api: Add position_type, strategy, apr_pct, source fields to
   RuntimePositionDto (needed by position detail popup).

5. api: Add tws_address: Option<String> to Metrics (needed by status bar).

6. tui config: Add split_pane: bool field and validate() method to TuiConfig.
   Add SPLIT_PANE env var override.

7. FIX scenarios.rs: cast strike_width filter comparison (f64 != u64).
8. FIX yield_curve.rs: borrow of moved symbol (clone in Row::new arms),
   reference in replace() call.
9. FIX ui/mod.rs hint bar: "1-8" -> "1-9" (was pre-existing bug).

Former-commit-id: 1dacfa210bb331fd293ab50e5faefbe4a93e6397

+++ agents/backend/services/tui_service/src/ui/scenarios.rs
@@ -76,7 +76,7 @@ pub fn filtered_scenarios(app: &App) -> Vec<ScenarioDto> {
         .into_iter()
         .filter(|s| {
             if let Some(w) = app.scenarios_strike_width_filter {
-                if s.strike_width != w {
+                if s.strike_width != w as f64 {
                     return false;
                 }
             }
